// src/_workspace/types/item-master/Vendor.ts
export interface VendorI {
  VENDOR_ID: number
  VENDOR_NAME: string
  VENDOR_ALPHABET: string
  ITEM_IMPORT_TYPE_ID: number
  ITEM_IMPORT_TYPE_NAME: string
  VENDOR_CD_PRONES: string
  VENDOR_NAME_PRONES: string
  INUSE: number
  MODIFIED_DATE: string
  UPDATE_BY: string
  No: number
}

export interface ItemImportTypeI {
  ITEM_IMPORT_TYPE_ID: number
  ITEM_IMPORT_TYPE_NAME: string
}